Refactored Router
