+++ src/router/base.js
@@ -1,0 +1,10 @@
+export default {
+  path: '/',
+  component: require('@/App.vue').default,
+  meta: {
+    main: true,
+    label: 'Home',
+    icon: require('@/assets/icons/home.svg').default
+  },
+  children: []
+}
