+++ src/router/index.js
@@ -1,28 +1,20 @@
 import Vue from 'vue'
 import Router from 'vue-router'
+
 import store from '@/store'
-import userRoutes from '@/router/UserRoutes'
-import adminRoutes from '@/router/AdminRoutes'
-import SharedRoutes from '@/router/SharedRoutes'
+import adminRoutes from '@/router/admin'
+import authRoutes from '@/router/auth'
+import userRoutes from '@/router/user'
+import appRoute from '@/router/base'
+
 import { userHasRole } from '@/auth'
 
 let router = new Router({
   mode: 'history',
-  routes: [...SharedRoutes]
+  routes: [...authRoutes]
 })
 
 export async function loadRoutes(user) {
-  let appRoute = {
-    path: '/',
-    component: require('@/App.vue').default,
-    meta: {
-      main: true,
-      label: 'Home',
-      icon: require('@/assets/icons/home.svg').default
-    },
-    children: []
-  }
-
   let routes = userRoutes
 
   if (userHasRole(user, 'ADMIN')) {
@@ -38,7 +30,7 @@ export async function loadRoutes(user) {
 export function reloadRouter() {
   Vue.router = new Router({
     mode: 'history',
-    routes: [...SharedRoutes]
+    routes: [...authRoutes]
   })
 }
